Inline reassembly test into reassemble.lua
